+++ init/venv-airflow-init.sh
@@ -1,8 +1,12 @@
 # Set versions to install into Airflow virtual environment
 AIRFLOW_VERSION=2.6.1
 
-EA_AIRFLOW_UTIL=0.2.0
-EDU_EDFI_AIRFLOW=0.2.0
+EA_AIRFLOW_UTIL_VERSION=0.2.0
+EDFI_API_CLIENT_VERSION=0.2.0
+EDU_EDFI_AIRFLOW_VERSION=0.2.0
+
+EARTHMOVER_VERSION=0.1.6
+LIGHTBEAM_VERSION=0.0.8
 
 
 ### set up airflow environment ###
@@ -21,9 +25,9 @@ pip install "apache-airflow[amazon, snowflake, slack, postgres, ssh, sftp]==${AI
 pip install airflow-dbt
 pip install pysftp
 
-pip install edfi_api_client
-pip install earthmover
-pip install lightbeam
+pip install edfi_api_client=="${EDFI_API_CLIENT_VERSION}"
+pip install earthmover=="${EARTHMOVER_VERSION}"
+pip install lightbeam=="${LIGHTBEAM_VERSION}"
 
 ### Install EDU packages as executable repos in the `code` directory.
 cd ~/code
@@ -35,8 +39,11 @@ pip install -e ea_airflow_util
 pip install -e edu_edfi_airflow
 
 # Checkout the most recent tagged releases.
-git -C ea_airflow_util  checkout "tags/v${EA_AIRFLOW_UTIL}"
-git -C edu_edfi_airflow checkout "tags/v${EDU_EDFI_AIRFLOW}"
+git -C ea_airflow_util  checkout "tags/v${EA_AIRFLOW_UTIL_VERSION}"
+git -C edu_edfi_airflow checkout "tags/v${EDU_EDFI_AIRFLOW_VERSION}"
 
 # Return to the original path.
 cd -
+
+
+# Add your custom Airflow environment variable overwrites here: